quest 04

+++ sqrt.go
@@ -2,6 +2,6 @@ package piscine
 
 import "math"
 
-func Sqrt(int nb) int {
+func Sqrt(int nb) float64 {
 	return math.Sqrt(float64(nb))
 }
